mqtt/requirements.txt
flask
flask-swagger-ui
paho-mqtt
